Fixed error for max RPM being a byte rather than an int and causing odd things..

+++ main/java/minecrafttransportsimulator/dataclasses/PackPartObject.java
@@ -15,7 +15,7 @@ public class PackPartObject{
     	public boolean isAutomatic;
     	public byte starterPower;
     	public byte starterDuration;
-    	public byte maxRPM;
+    	public int maxRPM;
     	public float fuelConsumption;
     	public float[] gearRatios;
     	public String[] repairMaterials;
